+++ client/views/cards/helpers.js
@@ -1,9 +1,1 @@
 Template.addCardForm.helpers({});
-
-Template.cardDetailWindow.helpers({
-
-    // live card 
-    card: function() {
-        return Cards.findOne(this.cardId);      
-    }
-})