+++ client/views/main/events.js
@@ -3,17 +3,3 @@ Template.pop.events({
         Utils.Pop.close();
     }
 });
-
-Template.windowOverlay.events({
-    'click .window-overlay': function(event, t) {
-        Utils.Overlay.close();
-        Utils.Pop.close();
-    },
-    'click .js-close-window': function(event) {
-        Utils.Overlay.close();
-        Utils.Pop.close();
-    },
-    'click .js-tab-parent': function(event, t) {
-        event.stopPropagation();
-    }
-});