+++ server/publications/cards.js
@@ -1,0 +1,7 @@
+Meteor.publishComposite('card', function(cardId) {
+    return {
+        find: function() {
+            return Cards.find({ _id: cardId });      
+        }
+    }
+});